Refactored button component, updated icons, and implemented dark theme enhancements
