+++ src/App.tsx
@@ -3,7 +3,7 @@ import { Footer } from './components/Footer';
 import { Header } from './components/Header';
 import { useAppSelector } from './hooks/hooks';
 import { Toaster } from 'react-hot-toast';
-import toastStyles from './components/Card/Toaster/toastStyles.module.scss';
+import toastStyles from './components/Actions/Toaster/toastStyles.module.scss';
 export const App = () => {
   const theme = useAppSelector((state) => state.theme.theme);
