+++ src/components/ProductDescription/About/About.tsx
@@ -1,7 +1,9 @@
 import { FC } from 'react';
 import styles from './About.module.scss';
+import cn from 'classnames';
 
 import { ProductSpec } from '../../../types/ProductSpec';
+import { useAppSelector } from '../../../hooks/hooks';
 
 type Props = {
   currentProductSpec: ProductSpec;
@@ -11,8 +13,11 @@ export const About: FC<Props> = (props) => {
   const {
     currentProductSpec: { description },
   } = props;
+
+  const theme = useAppSelector((state) => state.theme.theme);
+
   return (
-    <section className={styles.section}>
+    <section className={cn(styles.section, styles[theme])}>
       <h3 className={styles.section__title}>About</h3>
       {description.map((declaration, idx) => {
         return (
