+++ src/components/ProductDescription/Offer/Offer.tsx
@@ -3,19 +3,7 @@ import { NavLink, useLocation } from 'react-router-dom';
 import { ProductSpec } from '../../../types/ProductSpec';
 import styles from './Offer.module.scss';
 import cn from 'classnames';
-import buttonstyles from '../../Card/Card.module.scss';
-import favourites_icon from '/img/icons/Favourite.svg';
-import favourites_icon_white from '/img/icons/FavoriteWhite.svg';
-import favourites_icon_filled from '/img/icons/FavouritesFilled.svg';
-import { ButtonText } from '../../../types/ButtonText';
 import { Product } from '../../../types/Product';
-import { toast } from 'react-hot-toast';
-import { useDispatch } from 'react-redux';
-import { actions as favoritesActions } from '../../../features/favoritesProducts';
-import { increaseQuantity } from '../../../features/cartReducer';
-import { useAppSelector } from '../../../hooks/hooks';
-import { Theme } from '../../../types/Theme';
-import { ProductInCart } from '../../../types/ProductInCart';
 
 import {
   browserSupportedColors,
@@ -23,13 +11,8 @@ import {
   sortStrings,
   createNewItemId,
 } from '../../../utils';
-
-function isProductInList<T extends { id: string | number }>(
-  products: T[],
-  product: T
-): boolean {
-  return products.some((item) => item.id === product.id);
-}
+import { Actions } from '../../Actions';
+import { useAppSelector } from '../../../hooks/hooks';
 
 type Props = {
   currentProductSpec: ProductSpec;
@@ -51,53 +34,9 @@ export const Offer: FC<Props> = ({ currentProductSpec, product }) => {
     color,
   } = currentProductSpec;
 
-  const dispatch = useDispatch();
-  const favorites = useAppSelector(
-    (state) => state.favoritesProducts.favoritesProducts
-  );
-
-  const addFavorite = (product: Product) =>
-    dispatch(favoritesActions.add(product));
-  const removeFavorite = (product: Product) =>
-    dispatch(favoritesActions.remove(product));
-
-  const added = useAppSelector((state) => state.cartProducts.cartProducts);
-
-  const addToCart = (product: ProductInCart) =>
-    dispatch(increaseQuantity(product as ProductInCart));
-
-  const addToFavorite = () => {
-    if (isProductInList(favorites, product)) {
-      removeFavorite(product);
-      toast('Removed from favorites!', {
-        icon: '💔',
-      });
-    } else {
-      addFavorite(product);
-      toast('Added to favorites!', {
-        icon: '❤️',
-      });
-    }
-  };
-
-  const onAddToCart = () => {
-    if (!isProductInList(added, product as ProductInCart)) {
-      addToCart(product as ProductInCart);
-      toast('Added to cart!', {
-        icon: '🛒',
-      });
-    } else {
-      toast('Already in the cart!', {
-        icon: '🔔',
-      });
-    }
-  };
-
-  const theme = useAppSelector((state) => state.theme.theme);
-
   const location = useLocation();
-
   const category = location.pathname.split('/')[1];
+  const theme = useAppSelector((state) => state.theme.theme);
 
   const setColorsClasses = ({ isActive }: { isActive: boolean }) => {
     return cn({
@@ -109,7 +48,7 @@ export const Offer: FC<Props> = ({ currentProductSpec, product }) => {
   const setCapacityClasses = ({ isActive }: { isActive: boolean }) => {
     return cn({
       [styles.capacity__selector__item]: true,
-      [styles.capacity__selector__item_active]: isActive,
+      [styles.capacity__selector__item_inactive]: !isActive,
     });
   };
 
@@ -121,7 +60,7 @@ export const Offer: FC<Props> = ({ currentProductSpec, product }) => {
   };
 
   return (
-    <section className={styles.section}>
+    <section className={cn(styles.section, styles[theme])}>
       <div className={styles.colors}>
         <p className={styles.section__subtitle}>Available colors</p>
         <div className={styles.colors__selector}>
@@ -172,42 +111,7 @@ export const Offer: FC<Props> = ({ currentProductSpec, product }) => {
         <p className={styles.prices__regular}>${priceRegular}</p>
       </div>
 
-      <div className={buttonstyles.product_card__actions}>
-        <button
-          onClick={onAddToCart}
-          className={cn(buttonstyles.product_card__buy, {
-            [buttonstyles.product_card__added_to_cart]: isProductInList(
-              added,
-              product as ProductInCart
-            ),
-          })}
-        >
-          {isProductInList(added, product as ProductInCart)
-            ? ButtonText.ADDED
-            : ButtonText.ADD_TO_CART}
-        </button>
-
-        <button
-          onClick={addToFavorite}
-          className={cn(buttonstyles.product_card__favourite_button, {
-            [buttonstyles.product_card__added_to_favourite]: isProductInList(
-              favorites,
-              product
-            ),
-          })}
-        >
-          <img
-            src={
-              isProductInList(favorites, product)
-                ? favourites_icon_filled
-                : theme === Theme.DARK
-                  ? favourites_icon_white
-                  : favourites_icon
-            }
-            alt='favourite icon'
-          />
-        </button>
-      </div>
+      <Actions product={product}></Actions>
 
       <div className={styles.properties}>
         <ul className={styles.properties__list}>
