+++ src/firebase/AuthButton.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { FaRegUser } from 'react-icons/fa';
 import { useNavigate } from 'react-router-dom';
+import { useAppSelector } from '../hooks/hooks';
+import { Theme } from '../types/Theme';
 
 const AuthButton: React.FC = () => {
   const navigate = useNavigate();
@@ -9,10 +11,16 @@ const AuthButton: React.FC = () => {
     navigate('/auth');
   };
 
+  const theme = useAppSelector((state) => state.theme.theme);
+
   return (
     <button
       onClick={handleAuthRedirect}
-      style={{ background: 'none', border: 'none' }}
+      style={
+        theme === Theme.DARK
+          ? { background: 'none', border: 'none', color: 'white' }
+          : { background: 'none', border: 'none' }
+      }
     >
       <FaRegUser size={20} />
     </button>
