+++ src/pages/HomePage/HomePage.module.scss
@@ -29,7 +29,6 @@
   font-size: 32px;
   line-height: 41px;
   letter-spacing: -1%;
-  color: $primary;
 
   @include mixins.on-tablet {
     font-size: 48px;
